requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# WeatherBot/main.py: 10
Pillow == 8.2.0

# WeatherBot/lib/waveshare_epd/epd2in13_V2.py: 33
numpy == 1.19.5

# WeatherBot/main.py: 5
requests == 2.23.0

# WeatherBot/main.py: 6
schedule == 1.1.0
